+++ nupic/research/frameworks/dendrites/modules/__init__.py
@@ -22,6 +22,7 @@
 from .apply_dendrites import *
 from .boosted_dendrites import *
 from .dendrite_segments import DendriteSegments
+from .dendritic_mlp import DendriticMLP
 from .dendritic_layers import (
     AbsoluteMaxGatingDendriticLayer,
     AbsoluteMaxGatingDendriticLayer2d,
